アカウント取得のurl変更
